+++ nwt-cinema-movies/src/main/java/ba/unsa/etf/nwtcinemamovies/controllers/RolesController.java
@@ -29,7 +29,7 @@ public class RolesController extends AbstractController<RoleService> {
 
 	@Transactional
 	@RequestMapping(value = "{roleId}", method = RequestMethod.GET)
-	public String findById(@PathVariable("roleId") final Integer roleId) {
+	public String findById(@PathVariable("roleId") final Long roleId) {
 		return JSONConverter.toJSON(service.findById(Role.class, roleId));
 	}
